feat: Enable downloading of bulk user creation result
This commit enables the downloading of the result file after bulk user creation. The `bulkCreateUsers` method in the `UserService` now accepts a `responseType` option of `'blob'` to receive the response as a blob. The `UserMaintenanceComponent` has been updated to handle the response by calling the new `downloadFile` method, which creates a blob from the response data and triggers a file download.